Implemented optimistic state updates for adding and deleting wages. This will improve the overall flow of data within the app and limit reads and writes to the database unless completely necessary. Also prevents any de-syncing issues with the DB provided either of those operations should fail. Adjusted logic in shift items to accomodate the previous change. Added some slow down when exceeding the deletion threshold on swipe-to-delete.

+++ timekeeper/src/context/WagesContext.tsx
@@ -14,14 +14,21 @@ import {
   CreateWage,
   DeleteWage,
 } from "../database/database";
+import { Timestamp } from "firebase/firestore";
 import { useAuthContext } from "./AuthContext";
+import { v4 as uuidv4 } from "uuid";
+import { convertToCreateWageProps } from "../utils/convertData";
 
 interface WagesContextType {
   wages: WageObjectProps[];
   isLoadingWages: boolean;
   refreshWages: () => Promise<void>;
-  addWage: (wage: CreateWageProps) => Promise<void>;
-  deleteWage: (docId: string) => Promise<void>;
+  addWage: (
+    wage: CreateWageProps
+  ) => Promise<{ success: boolean; msg: string }>;
+  deleteWage: (docId: string) => Promise<boolean>;
+  setDateRange: (dateRangeQuery: DateRangeQuery) => void;
+  undoDelete: () => Promise<boolean>;
 }
 
 const WagesContext = createContext<WagesContextType | undefined>(undefined);
@@ -34,9 +41,14 @@ export const useWages = (): WagesContextType => {
   return context;
 };
 
+const sortWages = (wages: WageObjectProps[]) => {
+  return wages.sort((a, b) => b.shiftDate.toMillis() - a.shiftDate.toMillis());
+};
+
 export const WagesProvider = ({ children }: { children: ReactNode }) => {
   const { user } = useAuthContext();
   const [isLoadingWages, setIsLoadingWages] = useState<boolean>(false);
+  const [lastDeletedWage, setLastDeletedWage] = useState<WageObjectProps>();
   const [dateRangeQuery, setDateRangeQuery] = useState<DateRangeQuery>({
     queryType: "currentWeek",
   });
@@ -61,32 +73,82 @@ export const WagesProvider = ({ children }: { children: ReactNode }) => {
     getWages();
   }, [getWages]);
 
+  interface WageWithDocId extends Omit<CreateWageProps, "shiftDate"> {
+    docId: string;
+    shiftDate: Timestamp;
+  }
+
   const addWage = async (wage: CreateWageProps) => {
+    const tempId = uuidv4();
+    const tempWage: WageWithDocId = {
+      docId: tempId,
+      ...wage,
+      shiftDate: Timestamp.fromDate(wage.shiftDate),
+    };
+    // Optimistically update the state.
+    setWages((prevWages) => sortWages([...prevWages, tempWage]));
+
     try {
       const newWage = await CreateWage(user, wage);
-      if (newWage) {
-        setWages((prevWages) => [...prevWages, newWage]);
-      } else {
+      if (!newWage) {
         throw new Error(
-          "Wage was not successfully added to the database. Cancelling local state update..."
+          "Wage was not successfully added to the database. Rolling back local state update..."
         );
+      } else {
+        setWages((prevWages) =>
+          sortWages(
+            prevWages.map((prevWage) =>
+              prevWage.docId === tempId ? newWage : prevWage
+            )
+          )
+        );
+        return { success: true, msg: "Save successful" };
       }
     } catch (err) {
-      console.log(err);
+      setWages((prevWages) =>
+        sortWages(prevWages.filter((prevWage) => prevWage.docId !== tempId))
+      );
+
+      return { success: false, msg: "Operation to save data failed." };
     }
   };
 
   const deleteWage = async (docId: string) => {
     if (user) {
+      const wageToBeDeleted = wages.find((wage) => wage.docId === docId);
+      if (!wageToBeDeleted) {
+        console.log(`No wage found with docId: ${docId}`);
+        return false;
+      }
+
+      setLastDeletedWage(wageToBeDeleted);
+      setWages((prevWages) => prevWages.filter((wage) => wage.docId !== docId));
+
       try {
         await DeleteWage(user, docId);
-        setWages((prevWages) =>
-          prevWages.filter((wage) => wage.docId !== docId)
-        );
+        return true;
       } catch (err) {
-        console.log(err);
+        setWages((prevWages) => sortWages([...prevWages, wageToBeDeleted]));
+        setLastDeletedWage(undefined);
+        return false;
       }
     }
+    return false;
+  };
+
+  const undoDelete = async () => {
+    if (lastDeletedWage !== undefined && user !== null) {
+      const convertedWage = convertToCreateWageProps(lastDeletedWage);
+      const response = await addWage(convertedWage);
+      if (response.success) {
+        setLastDeletedWage(undefined);
+        return true;
+      } else {
+        return false;
+      }
+    }
+
+    return false;
   };
 
   const setDateRange = (newDateRangeQuery: DateRangeQuery) => {
@@ -100,6 +162,7 @@ export const WagesProvider = ({ children }: { children: ReactNode }) => {
     addWage,
     deleteWage,
     setDateRange,
+    undoDelete,
   };
 
   return (
